Move startExpressServer to its own file
